Refactor: unificación de rutas administrativas en admin.routes.js
