+++ frontend/src/App.jsx
@@ -4,6 +4,7 @@ import CertificatesRoute from "./routes/CertificatesRoute";
 import EventsRoute from "./routes/EventsRoute";
 import Register from "./views/Register";
 import MyInscriptions from "./views/MyInscriptions";
+import AdminEventInscription from "./views/admin/AdminEventInscription";
 
 // Importación de ToastContainer y estilos para mostrar notificaciones tipo toast
 import { ToastContainer } from "react-toastify";
@@ -32,6 +33,12 @@ function App() {
 
           {/* Página de inscripciones */}
           <Route path="/inscripciones" element={<MyInscriptions />} />
+
+          {/* Página de administración de inscripciones para eventos */}
+          <Route
+            path="/admin/eventos/:id/inscripciones"
+            element={<AdminEventInscription />}
+          />
         </Routes>
 
         {/* Contenedor global para notificaciones tipo toast */}
